requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-crispy-forms==1.14.0
django-google-maps==0.13.0
django-model-utils==4.2.0
django-tinymce==3.4.0
djangorestframework==3.13.1
Pillow==9.2.0
pytz==2022.1
sqlparse==0.4.2
